move getCalendarLinks
