+++ packages/platform/libraries/index.ts
@@ -120,7 +120,7 @@ export { getTeamMemberEmailForResponseOrContactUsingUrlQuery };
 export { SelectedCalendarRepository } from "@calcom/lib/server/repository/selectedCalendar";
 export { encryptServiceAccountKey } from "@calcom/lib/server/serviceAccountKey";
 export { createHandler as createApiKeyHandler } from "@calcom/trpc/server/routers/viewer/apiKeys/create.handler";
-export { getCalendarLinks } from "@calcom/lib/bookings/getCalendarLinks";
+export { getCalendarLinks } from "@calcom/features/bookings/lib/getCalendarLinks";
 
 export { findTeamMembersMatchingAttributeLogic } from "@calcom/lib/raqb/findTeamMembersMatchingAttributeLogic";
 export type { TFindTeamMembersMatchingAttributeLogicInputSchema } from "@calcom/trpc/server/routers/viewer/attributes/findTeamMembersMatchingAttributeLogic.schema";
